pkp/pkp-lib#9034 Incorporated the I7167_RemoveDuplicatedUserSettingsAndDeprecatedFields into the pre-flight migration to avoid false positives at the MergeLocalesMigration checks
