+++ classes/migration/upgrade/v3_4_0/PreflightCheckMigration.php
@@ -20,7 +20,6 @@ use APP\statistics\StatisticsHelper;
 use Exception;
 use Illuminate\Database\MySqlConnection;
 use Illuminate\Database\PostgresConnection;
-use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Schema;
 use PKP\db\DAORegistry;
@@ -54,6 +53,8 @@ abstract class PreflightCheckMigration extends \PKP\migration\Migration
                 throw new Exception("There are one or more log files that were unable to finish processing. This happens when the scheduled task to process usage stats logs encounters a failure of some kind. These logs must be repaired and reprocessed or removed before the upgrade can continue. The logs can be found in the folders reject, processing and stage in {$usageStatsDir}.");
             }
 
+            // It's needed to clear the duplicated settings before looking for duplicated localized data to avoid false positives
+            $this->clearDuplicatedUserSettings();
             // email_templates_default_data keys should not conflict after locale migration
             // See MergeLocalesMigration (#8598)
             $affectedLocales = MergeLocalesMigration::getAffectedLocales();
@@ -748,6 +749,76 @@ abstract class PreflightCheckMigration extends \PKP\migration\Migration
                 'Contact name or email is missing for context(s) with path(s) [%s]. Please set those before upgrading.',
                 $missingContactContexts->pluck('path')->implode(',')
             )
+        );
+    }
+
+    /**
+     * Clears duplicated user_settings
+     * This method used to be a migration, it has been incorporated at the pre-flight to avoid issues with the checks introduced by the MergeLocalesMigration
+     * Given that it operates on duplicated entries, it should be ok to run it several times
+     * @see https://github.com/pkp/pkp-lib/issues/7167
+     */
+    protected function clearDuplicatedUserSettings(): void
+    {
+        // Locates and removes duplicated user_settings
+        // The latest code stores settings using assoc_id = 0 and assoc_type = 0. Which means entries using null or anything else are outdated.
+        // Note: Old versions (e.g. OJS <= 2.x) made use of these fields to store some settings, but they have been removed years ago, which means they are safe to be discarded.
+        if (DB::connection() instanceof PostgresConnection) {
+            DB::unprepared(
+                "DELETE FROM user_settings s
+                USING user_settings duplicated
+                -- Attempts to find a better fitting record among the duplicates (preference is given to the smaller assoc_id/assoc_type values)
+                LEFT JOIN user_settings best
+                    ON best.setting_name = duplicated.setting_name
+                    AND best.user_id = duplicated.user_id
+                    AND best.locale = duplicated.locale
+                    AND (
+                        COALESCE(best.assoc_id, 999999) < COALESCE(duplicated.assoc_id, 999999)
+                        OR (
+                            COALESCE(best.assoc_id, 999999) = COALESCE(duplicated.assoc_id, 999999)
+                            AND COALESCE(best.assoc_type, 999999) < COALESCE(duplicated.assoc_type, 999999)
+                        )
+                    )
+                -- Locates all duplicated settings (same key fields, except the assoc_type/assoc_id)
+                WHERE s.setting_name = duplicated.setting_name
+                    AND s.user_id = duplicated.user_id
+                    AND s.locale = duplicated.locale
+                    AND (
+                        COALESCE(s.assoc_type, -999999) <> COALESCE(duplicated.assoc_type, -999999)
+                        OR COALESCE(s.assoc_id, -999999) <> COALESCE(duplicated.assoc_id, -999999)
+                    )
+                    -- Ensures a better record was found (if not found, it means the current duplicated record is the best and shouldn't be removed)
+                    AND best.user_id IS NOT NULL"
+            );
+            return;
+        }
+
+        DB::unprepared(
+            "DELETE s
+            FROM user_settings s
+            -- Locates all duplicated settings (same key fields, except the assoc_type/assoc_id)
+            INNER JOIN user_settings duplicated
+                ON s.setting_name = duplicated.setting_name
+                AND s.user_id = duplicated.user_id
+                AND s.locale = duplicated.locale
+                AND (
+                    COALESCE(s.assoc_type, -999999) <> COALESCE(duplicated.assoc_type, -999999)
+                    OR COALESCE(s.assoc_id, -999999) <> COALESCE(duplicated.assoc_id, -999999)
+                )
+            -- Attempts to find a better fitting record among the duplicates (preference is given to the smaller assoc_id/assoc_type values)
+            LEFT JOIN user_settings best
+                ON best.setting_name = duplicated.setting_name
+                AND best.user_id = duplicated.user_id
+                AND best.locale = duplicated.locale
+                AND (
+                    COALESCE(best.assoc_id, 999999) < COALESCE(duplicated.assoc_id, 999999)
+                    OR (
+                        COALESCE(best.assoc_id, 999999) = COALESCE(duplicated.assoc_id, 999999)
+                        AND COALESCE(best.assoc_type, 999999) < COALESCE(duplicated.assoc_type, 999999)
+                    )
+                )
+            -- Ensures a better record was found (if not found, it means the current duplicated record is the best and shouldn't be removed)
+            WHERE best.user_id IS NOT NULL"
         );
     }
 }
